Fixed generator

+++ priority-research/generator.cpp
@@ -4,7 +4,7 @@ typedef long long ll;
 
 int main(int argc, char** argv) {
     registerGen(argc, argv, 1);
-    int n = opt<int>("n");
+    ll n = opt<ll>("n");
 
     cout << n << "\n";
